Extracted two more files - model_params to keep class with model parameters and plotting_functions to keep all plotting functions in one place
